access for content

+++ server/src/services/user.service.js
@@ -1,5 +1,6 @@
 const {User, Action} = require("../models");
-const {UserEnum: {EUserStatus}} = require("../enums");
+const {UserEnum: {EUserStatus, EUserRole}} = require("../enums");
+const {ApiError} = require("../errors");
 
 class UserService {
     async findAll() {
@@ -15,10 +16,23 @@ class UserService {
     }
 
     async banById(id) {
-        await Promise.all([
-            Action.deleteMany({_userId: id}),
-            User.updateOne({_id: id}, {status: EUserStatus.BANNED})
-        ])
+        try {
+            const userToBan = await User.findById(id);
+            if (userToBan.role === EUserRole.ADMIN) {
+                throw new ApiError("You can not ban another admin", 400)
+            }
+
+            await Promise.all([
+                Action.deleteMany({_userId: id}),
+                User.updateOne({_id: id}, {status: EUserStatus.BANNED})
+            ])
+        } catch (e) {
+            throw new ApiError(e.message, e.status)
+        }
+    }
+
+    async setAdmin(id) {
+        await User.updateOne({_id: id}, {role: EUserRole.ADMIN});
     }
 }
